requirements.txt
boto3>=1.35.0
nova-act>=0.1.0
streamlit>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
asyncio
aiohttp>=3.9.0
numpy>=1.26.0
faiss-cpu>=1.8.0
PyPDF2>=3.0.0
pillow>=10.0.0
pyaudio>=0.2.14
sounddevice>=0.4.6
soundfile>=0.12.1
requests>=2.31.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
